Added working functionality to most of the single matrix functions

+++ adjoint.h
@@ -1,1 +1,1 @@
-double **adjoint(int n, int **matrix);
+double **adjoint(int n, double **matrix);
